Refactored db/request.js to db/requests.js for consistency
